requirements.txt
requests
pandas
pandas_ta
pyTelegramBotAPI
python-dotenv
